pyproject.toml
[tool.poetry]
name = "fampay"
version = "0.1.0"
description = ""
authors = ["s-arvind <arvindsharma1317@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
google-api-python-client = "^2.14.0"
python-dateutil = "^2.8.2"
fastapi = "^0.66.1"
elasticsearch = {extras = ["async"], version = "^7.13.4"}
gunicorn = "^20.1.0"
python-decouple = "^3.4"
pymongo = "^3.12.0"
uvicorn = "^0.14.0"
loguru = "^0.5.3"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
